CS48007-IoT_Sensing_Systems/PA3-Remote_Sensing_System/server/requirements.txt
flask==2.2.2
flask-Cors==3.0.10
ipython==8.5.0
keras==2.7.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.23.4
opencv-contrib-python==4.6.0.66
waitress
pandas==1.5.0
parso==0.8.3
psutil==5.9.3
pickleshare==0.7.5
Pillow==9.2.0
PyYAML==6.0
pyzmq==24.0.1
requests==2.26.0
scipy==1.9.2
seaborn==0.12.1
six==1.16.0
tf==1.0.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
